Move a few more tests
