+++ packages/fs-detectors/test/fixtures/01-zero-config-api/package.json
