+++ packages/fs-detectors/test/fixtures/02-zero-config-api/.gitignore
@@ -1,0 +1,1 @@
+now.json
