+++ packages/fs-detectors/test/fixtures/02-zero-config-api/api/date/index.js
@@ -1,0 +1,3 @@
+module.exports = (req, res) => {
+  res.end('hello from api/date/index.js');
+};
