+++ packages/fs-detectors/test/fixtures/02-zero-config-api/package.json
@@ -1,0 +1,5 @@
+{
+  "scripts": {
+    "build": "mkdir -p public && echo 'hello from index.txt' > public/index.txt"
+  }
+}
